refactor: move int grid image creation to a new impl module for Definitions
